Cloudflare Turnstile & PayOS 2 - Upgraded payos-java dependency to version 2.0.1. - Fixed payos payment link - Email localization - Migrated from Google RECAPTCHA to Cloudflare Turnstile

+++ back-end/src/main/java/com/ring/model/entity/PaymentInfo.java
@@ -20,16 +20,8 @@ public class PaymentInfo {
 
     @Id
     @Column(nullable = false, updatable = false)
-    @SequenceGenerator(
-            name = "primary_sequence",
-            sequenceName = "primary_sequence",
-            allocationSize = 1,
-            initialValue = 10000
-    )
-    @GeneratedValue(
-            strategy = GenerationType.SEQUENCE,
-            generator = "primary_sequence"
-    )
+    @SequenceGenerator(name = "primary_sequence", sequenceName = "primary_sequence", allocationSize = 1, initialValue = 10000)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "primary_sequence")
     private Long id;
 
     @Enumerated(EnumType.STRING)
@@ -44,7 +36,7 @@ public class PaymentInfo {
     private String checkoutUrl;
 
     @Column
-    private Integer amount;
+    private Long amount;
 
     @Column(length = 500)
     private String description;
